refactor: split out code for view model
